Add timeout configuration to environment settings; refactor service imports
