+++ expensevault.client/.env.production
@@ -2,3 +2,4 @@ VITE_REACT_APP_USE_MOCK_DATA=false
 VITE_API_URL=https://api.expensevault.com
 VITE_ENV=production
 VITE_CONSOLE_LOG=false
+VITE_TIMEOUT=30000
