+++ expensevault.client/.env.template
@@ -2,3 +2,4 @@ VITE_REACT_APP_USE_MOCK_DATA=false
 VITE_API_URL=https://localhost:7030
 VITE_ENV=development
 VITE_CONSOLE_LOG=false
+VITE_TIMEOUT=30000
