+++ expensevault.client/src/api/client.test.ts
@@ -1,8 +1,11 @@
 import MockAdapter from 'axios-mock-adapter';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { httpClient, unauthenticatedHttpClient } from './http.service';
-import { httpServiceGet } from './http.service';
+import {
+  httpClient,
+  httpServiceGet,
+  unauthenticatedHttpClient,
+} from './client';
 
 // Mock the auth utilities
 vi.mock('../../../utils/auth-util', () => ({
