+++ expensevault.client/src/api/endpoints/auth.service.ts
@@ -1,15 +1,15 @@
-import { ApiResult } from '../../../shared/types/common';
+import { ApiResult } from '../../shared/types/common';
 import {
   LoginCommand,
   LoginResponse,
-} from '../../../shared/types/common/backend-model';
-import { httpServicePost } from '../axios/http.service';
+} from '../../shared/types/common/backend-model';
+import { httpServicePost } from '../client';
 
 export const Login = async ({
   username,
   password,
 }: LoginCommand): Promise<ApiResult<LoginResponse>> => {
-  const response = await httpServicePost('/api/auth/login', {
+  const response = await httpServicePost<LoginResponse>('/api/auth/login', {
     username,
     password,
   });
